+++ basic/tests/fixtures/AircraftTypeResourceFixture.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace tests\fixtures;
+
+use yii\test\ActiveFixture;
+
+class AircraftTypeResourceFixture extends ActiveFixture
+{
+    public $tableName = '{{%aircraft_type_resource}}';
+    public $depends   = ['tests\fixtures\AircraftTypeFixture'];
+}